move into football data specific handler and render files
