+++ dotcom-rendering/src/server/render.article.web.tsx
@@ -2,9 +2,7 @@ import { isString } from '@guardian/libs';
 import { ArticlePage } from '../components/ArticlePage';
 import { ConfigProvider } from '../components/ConfigContext';
 import { isAmpSupported } from '../components/Elements.amp';
-import { FootballDataPage } from '../components/FootballDataPage';
 import { LiveBlogRenderer } from '../components/LiveBlogRenderer';
-import type { FEFootballDataPage } from '../feFootballDataPage';
 import {
 	ArticleDesign,
 	type ArticleFormat,
@@ -22,10 +20,7 @@ import { renderToStringWithEmotion } from '../lib/emotion';
 import { getCurrentPillar } from '../lib/layoutHelpers';
 import { polyfillIO } from '../lib/polyfill.io';
 import { extractNAV } from '../model/extract-nav';
-import {
-	createGuardian,
-	createGuardian as createWindowGuardian,
-} from '../model/guardian';
+import { createGuardian as createWindowGuardian } from '../model/guardian';
 import type { Article } from '../types/article';
 import type { Config } from '../types/configContext';
 import type { FEElement } from '../types/content';
@@ -304,81 +299,4 @@ export const renderBlocks = ({
 	);
 
 	return `${extractedCss}${html}`;
-};
-
-export const renderFootballDataPage = (footballData: FEFootballDataPage) => {
-	const config: Config = {
-		renderingTarget: 'Web',
-		darkModeAvailable:
-			footballData.config.abTests.darkModeWebVariant === 'variant',
-		assetOrigin: ASSET_ORIGIN,
-		editionId: footballData.editionId,
-	};
-
-	const { html, extractedCss } = renderToStringWithEmotion(
-		<ConfigProvider value={config}>
-			<FootballDataPage footballData={footballData} />
-		</ConfigProvider>,
-	);
-
-	const build = getModulesBuild({
-		switches: footballData.config.switches,
-		tests: footballData.config.abTests,
-	});
-
-	/**
-	 * The highest priority scripts.
-	 * These scripts have a considerable impact on site performance.
-	 * Only scripts critical to application execution may go in here.
-	 * Please talk to the dotcom platform team before adding more.
-	 * Scripts will be executed in the order they appear in this array
-	 */
-	const prefetchScripts = [
-		polyfillIO,
-		getPathFromManifest(build, 'frameworks.js'),
-		getPathFromManifest(build, 'index.js'),
-		process.env.COMMERCIAL_BUNDLE_URL ??
-			footballData.config.commercialBundleUrl,
-	].filter(isString);
-	const legacyScripts = [
-		getPathFromManifest('client.web.legacy', 'frameworks.js'),
-		getPathFromManifest('client.web.legacy', 'index.js'),
-	];
-	const scriptTags = generateScriptTags([
-		...prefetchScripts,
-		...legacyScripts,
-	]);
-
-	const pageHtml = htmlPageTemplate({
-		scriptTags,
-		css: extractedCss,
-		html,
-		//title,
-		//description,
-		guardian: createGuardian({
-			editionId: footballData.editionId,
-			stage: footballData.config.stage,
-			frontendAssetsFullURL: footballData.config.frontendAssetsFullURL,
-			revisionNumber: footballData.config.revisionNumber,
-			sentryPublicApiKey: footballData.config.sentryPublicApiKey,
-			sentryHost: footballData.config.sentryHost,
-			keywordIds: footballData.config.keywordIds,
-			dfpAccountId: footballData.config.dfpAccountId,
-			adUnit: footballData.config.adUnit,
-			ajaxUrl: footballData.config.ajaxUrl,
-			googletagUrl: footballData.config.googletagUrl,
-			switches: footballData.config.switches,
-			abTests: footballData.config.abTests,
-			brazeApiKey: footballData.config.brazeApiKey,
-			isPaidContent: footballData.config.isPaidContent,
-			contentType: footballData.config.contentType,
-			googleRecaptchaSiteKey: footballData.config.googleRecaptchaSiteKey,
-		}),
-		keywords: footballData.config.keywords,
-		renderingTarget: 'Web',
-		weAreHiring: !!footballData.config.switches.weAreHiring,
-		config,
-	});
-
-	return { html: pageHtml, prefetchScripts };
 };
